aws-cost-agent-framework/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
asyncio
plotly>=5.17.0
pandas>=2.0.0
Pillow>=10.0.0
PyJWT>=2.8.0